Add the start of a testharness
In order to test grmphl we need a repeatable test setup on which
runs can be executed. Add a start of a harness setup with which
a upstream/fork repo structure is created and merged. There is
no support for automatically executing grmphl on the structure
yet as there is no non-interactive mode (yet).