Initial commit of remerge script
remerge.pl is intended to for replaying large merges that have been
worked on collaboratively in a branch while origin/master has been
allowed to progress. This script replays the merge as far as it can
with the exception that a set of commits can be excluded for later
manual addition to make the git history more readable.

This is a not very general purpose script but is intended for a
specific merge in mind. It will be cleaned up to be more generic at
some point.